backend/service-mqtt/requirements.txt
django
paho-mqtt